requirements.txt
alembic==0.8.10
apache-airflow==1.9.0
beautifulsoup4==4.6.0
bleach==2.1.2
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
croniter==0.3.23
dill==0.2.7.1
docutils==0.14
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.10
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.8.1
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
nltk==3.3
numpy==1.14.3
ordereddict==1.1
pandas==0.23.0
psutil==4.4.2
psycopg2==2.7.4
Pygments==2.2.0
python-daemon==2.1.2
python-dateutil==2.7.3
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2018.4
PyYAML==3.12
requests==2.18.4
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.2.7
tabulate==0.7.7
thrift==0.11.0
Unidecode==1.0.22
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
zope.deprecation==4.3.0
